Melhorias e proteção cadastro usuário
- Adicionado local nos index para exbir erros
- Exibe qtd de produto por categoria
- Não deixa apagar produto se tiver categorias associadas a ele
- Usuários podem apenas gerenciar, alterar o próprio cadastro.
- Ao alterar o proprio e-mail, ele valida se já existe outro não considerando ele mesmo.
- Campo de e-mail verificado é deixado com null quando o campo de email é alterado.

+++ PI3/app/Http/Controllers/CategoriesController.php
@@ -67,10 +67,22 @@ class CategoriesController extends Controller
     public function destroy($id)
     {
         $category = Category::withTrashed()->where('id', $id)->firstOrFail();
-        if($category->trashed()){
+
+        $qtdProdutos = $category->products()->count();
+
+        if( $qtdProdutos > 0 )
+        {
+            session()->flash('error', "Existem $qtdProdutos produtos com essa categoria!");
+            return redirect()->back();
+        }
+
+        if($category->trashed())
+        {
             $category->forceDelete();
             session()->flash('success', 'Categoria removida com sucesso!');
-        }else{
+        }
+        else
+        {
             $category->delete();
             session()->flash('success', 'Categoria movida para lixeira com sucesso!');
         }
